Pull payout code into separate module.
